requirements.txt
openai
python-dotenv
jieba
fastapi
uvicorn
pydantic
requests 